% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{countInGolden}
\alias{countInGolden}
\title{Counts features from DIA-NN report e.g. Precursor.Id in golden standard list}
\usage{
countInGolden(
  report,
  golden_standard,
  feature_var = "Precursor.Id",
  QC_type = "linear_peptides"
)
}
\arguments{
\item{QC_type}{comment, default (linear peptides)}

\item{\strong{report}}{DIA-NN report file}

\item{\strong{golden_standard}}{golden standard df with \code{feature_var}}

\item{\strong{feature_var}}{feature variable to cout}
}
\value{

}
\description{
Counts features from DIA-NN report e.g. Precursor.Id in golden standard list
}
\examples{
LIN_PROTOCOL = "5-min-sswath"
LIN_N = 10
LIN_COR = 0.8
golden_linear = linear_peptides_5minsswath \%>\% ungroup \%>\% filter(protocol == LIN_PROTOCOL, n >= LIN_N, cor >= LIN_COR)
countInGolden(report = diann_report,
             golden_standard = golden_linear,
             QC_type = paste(LIN_PROTOCOL, LIN_N, LIN_COR, sep = ":" )) -> golden_linear_counts

}
