% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{countStats}
\alias{countStats}
\title{Counts Z-score stats of precursors}
\usage{
countStats(report, Q_THR = 0.01, feature_value = "Precursor.Quantity")
}
\arguments{
\item{\strong{report}}{DIA-NN report file}

\item{\strong{Q_THR}}{Q-value filtering threshold, default: 0.01}

\item{\strong{feature_value}}{feature value to count}
}
\value{
returns tbl with Z-scores of feature values from \code{feature_value}
}
\description{
Counts Z-score stats of precursors
}
\examples{
diann_report \%>\%
 filter(!grepl(pattern = ".*?BLANK.*?", perl = T, ignore.case = T, x = File.Name)) \%>\% #removes blanks
        countStats() -> report_stats

}
