% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{create_metadata}
\alias{create_metadata}
\title{Creates metadata from DIA-NN report}
\usage{
create_metadata(
  report,
  INTO = c("Batch", "sample_id", "run_order", "plate_pos", "well"),
  file_pattern = "(.*?)_(.*?)_([0-9]+)_([0-9]+)_([A-Za-z0-9]+)\\\\.wiff\\\\.dia$"
)
}
\arguments{
\item{\strong{report}}{DIA-NN report file}

\item{\strong{INTO}}{extracted columns in metadata output file, e.g.
\code{c("Batch", "sample_id", "run_order", "plate_pos",  "well")}}

\item{\strong{file_pattern}}{regular expression for extracting experimental metadata, e.g.
\code{"(.*?)_(.*?)_([0-9]+)_([0-9]+)_([A-Za-z0-9]+)\\\\.wiff\\\\.dia$"}}
}
\value{
returns metadata tibble
}
\description{
Creates metadata from DIA-NN report
}
\examples{
create_metadata(diann_report,
                INTO = c("Batch", "sample_id", "run_order", "plate_pos",  "well"),
                file_pattern = "(.*?)_(.*?)_([0-9]+)_([0-9]+)_([A-Za-z0-9]+)\\\\.wiff\\\\.dia$")

}
