% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Counts prevalence of features (e.g. Precursor.Id) from DIA-NN report}
\usage{
getPrevalence(report, Q_THR = 0.01, feature_var = "Precursor.Id")
}
\arguments{
\item{\strong{report}}{DIA-NN report file}

\item{\strong{feature_var}}{a single feature such as Precursor.Id}

\item{\strong{Q_THR}}{Q-value filtering threshold, default: 0.01}
}
\value{
df with feature prevalence values
}
\description{
Counts prevalence of features (e.g. Precursor.Id) from DIA-NN report
}
\examples{
Precursor.Id_prevalence <- getPrevalence(diann_report, feature_var = "Precursor.Id")
}
