% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lappend}
\alias{lappend}
\title{Append List}
\usage{
lappend(lst, obj)
}
\arguments{
\item{lst}{a list to append object to}

\item{obj}{object to append}
}
\value{
a list with appended object
}
\description{
The function appends objects into given list.
}
\examples{
myList <- list()
myList <- lappend(myList, mtcars)
}
