% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotExperiment}
\alias{plotExperiment}
\title{Plots Experiment}
\usage{
plotExperiment(
  report,
  metadata,
  Q_THR = 0.01,
  feature_var = "Precursor.Id",
  feature_value = "Precursor.Quantity",
  run_order_var = "run_order",
  subtitle = "",
  qc_pattern = ".*?QC.*?"
)
}
\arguments{
\item{metadata}{experiment description to arrange by run_order output from \code{\link{create_metadata}}}

\item{run_order_var}{variable for sample running order}

\item{subtitle}{a string for a plot's subtitle}

\item{qc_pattern}{QC file pattern}

\item{\strong{report}}{DIA-NN report file}

\item{\strong{Q_THR}}{Q-value filtering threshold, default: 0.01}

\item{\strong{feature_value}}{feature value to plot, default: "Precursor.Quantity"}

\item{\strong{feature_var}}{feature variable, default: "Precursor.Id"}
}
\value{

}
\description{
Plots Experiment
}
\examples{
metadata <- create_metadata(diann_report)
plotExperiment(diann_report, metadata = metadata, Q_THR = 0.01)

}
