% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGoldenCounts}
\alias{plotGoldenCounts}
\title{Plots}
\usage{
plotGoldenCounts(
  data_counts,
  metadata,
  golden_thr = 0.9,
  qc_pattern = ".*?QC.*?"
)
}
\arguments{
\item{metadata}{experiment description to arrange by run_order output from \code{\link{create_metadata}}}

\item{golden_thr}{threshold for QC line, default = 0.8}

\item{qc_pattern}{QC file pattern}

\item{\strong{data_counts}}{output from \code{\link{create_metadata}}}
}
\value{
list of plot and samples bellow golden_thr
}
\description{
Plots
}
\examples{
LIN_PROTOCOL <- "5-min-sswath"
LIN_N <- 10
LIN_COR <- 0.8
LIN_PROTOCOL = "5-min-sswath"
golden_linear <- linear_peptides_5minsswath \%>\% ungroup \%>\% filter(protocol == LIN_PROTOCOL, n >= LIN_N, cor >= LIN_COR)

countInGolden(report = diann_report,
             golden_standard = golden_linear,
             QC_type = paste(LIN_PROTOCOL, LIN_N, LIN_COR, sep = ":" )) -> golden_linear_counts

metadata <- create_metadata(diann_report)
ret = plotGoldenCounts(golden_linear_counts, metadata, golden_thr = 0.1)

}
