% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotMissingness}
\alias{plotMissingness}
\title{Plots missingness values}
\usage{
plotMissingness(
  report,
  metadata,
  feature_var = "Precursor.Id",
  run_order_var = "run_order",
  subtitle = "",
  prevalence_filter = 0.1
)
}
\arguments{
\item{metadata}{experiment description to arrange by run_order output from \code{\link{create_metadata}}}

\item{run_order_var}{sample running order}

\item{subtitle}{string for a subtitle}

\item{prevalence_filter}{filter out non-prevalent proteins, default: 0.1}

\item{\strong{report}}{DIA-NN report file}

\item{\strong{feature_var}}{feature for plotting missingness values}
}
\value{

}
\description{
Plots missingness values
}
\examples{
metadata <- create_metadata(diann_report)
plotMissingness(diann_report, metadata = metadata, prevalence_filter = 0.5)
}
