% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPrecursors}
\alias{plotPrecursors}
\title{Plots features (e.g. Precursor.Id) across entire batch run}
\usage{
plotPrecursors(
  report,
  metadata,
  Q_THR,
  n = 6,
  run_order_var = "run_order",
  feature_var = "Precursor.Id",
  feature_value = "Precursor.Quantity",
  qc_pattern = ".*?QC.*?",
  subtitle = "TEST",
  seed = 123
)
}
\arguments{
\item{metadata}{experiment description to arrange by run_order output from \code{\link{create_metadata}}}

\item{qc_pattern}{QC file pattern, default: ".*?QC.*?"}

\item{subtitle}{a string for a plot's subtitle}

\item{\strong{report}}{DIA-NN report file}

\item{\strong{Q_THR}}{Q-value filtering threshold, default: 0.01}

\item{\strong{feature_value}}{feature value to plot, default: "Precursor.Quantity"}

\item{\strong{feature_var}}{feature variable, default: "Precursor.Id"}

\item{\strong{run_order_var}}{variable for sample running order, default: "run_order"}

\item{\string{n}}{number of features to plot, default: 6}
}
\value{
returns feature plot
}
\description{
Plots features (e.g. Precursor.Id) across entire batch run
}
\examples{
metadata <- create_metadata(diann_report)
plotPrecursors(diann_report, metadata = metadata, Q_THR = 0.01, n = 6, run_order_var = "run_order", feature_var="Precursor.Id", feature_value = "Precursor.Quantity", qc_pattern = ".*?MSQC.*?")
}
