% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_object}
\alias{save_object}
\title{Saves RData object}
\usage{
save_object(object, DATA_PATH, suffix = "")
}
\arguments{
\item{object}{an object to save}

\item{DATA_PATH}{a directory path to save object}

\item{suffix}{(optional, Default = "") adds suffix string to the saved object}
}
\value{

}
\description{
The function saves an object into specified directory naming with the name of the object.
}
\examples{
save_object(mtcars, DATA_PATH="./", suffix="myresults")
}
